feat(student): prevent assigning teacher as student and vice versa
Added validation in the Student model to ensure that a user who is assigned as a teacher cannot be assigned to a student. Similarly, a user who is already a student cannot be assigned as a teacher. If such a case occurs, an error message is shown in the admin panel.

Issue #11 #2